requirements.txt
asgiref==3.11.1
Django==6.0.3
django-redis==6.0.0
djangorestframework==3.17.0
redis==7.3.0
sqlparse==0.5.5
